move function unit test
